% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\name{Person}
\alias{Person}
\title{R6 Class Representing a Person}
\description{
A person has a name and a hair color.
}
\details{
A person can also greet you.
}
\examples{

## ------------------------------------------------
## Method `Person$set_hair`
## ------------------------------------------------

P <- Person("Ann", "black")
P$hair
P$set_hair("red")
P$hair
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{First or full name of the person.}

\item{\code{hair}}{Hair color of the person.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Person$new()}}
\item \href{#method-set_hair}{\code{Person$set_hair()}}
\item \href{#method-greet}{\code{Person$greet()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new person object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$new(name = NA, hair = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name.}

\item{\code{hair}}{Hair color.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Person` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_hair"></a>}}
\subsection{Method \code{set_hair()}}{
Change hair color.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$set_hair(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New hair color.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{P <- Person("Ann", "black")
P$hair
P$set_hair("red")
P$hair
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-greet"></a>}}
\subsection{Method \code{greet()}}{
Say hi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$greet()}\if{html}{\out{</div>}}
}

}
}
