% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6md.R
\name{R6md}
\alias{R6md}
\title{R6 Class Representing a markdown report}
\description{
A report contains chunks of code contributing to an R
Markdown report
}
\details{
Chunks can be added or removed from the report and the updated code will
trigger recreation of the stored markdown document.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.n_exp}}{total number of chunks}

\item{\code{.char_code_blocks}}{total number of chunks}

\item{\code{.n_exp}}{total number of chunks}

\item{\code{.n_exp}}{total number of chunks}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.n_exp}}{total number of chunks}

\item{\code{.char_code_blocks}}{total number of chunks}

\item{\code{.n_exp}}{total number of chunks}

\item{\code{.n_exp}}{total number of chunks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{R6md$new()}}
\item \href{#method-print}{\code{R6md$print()}}
\item \href{#method-print2}{\code{R6md$print2()}}
\item \href{#method-set_n_exp}{\code{R6md$set_n_exp()}}
\item \href{#method-deparse_code}{\code{R6md$deparse_code()}}
\item \href{#method-add_code}{\code{R6md$add_code()}}
\item \href{#method-add_code_block}{\code{R6md$add_code_block()}}
\item \href{#method-run_code}{\code{R6md$run_code()}}
\item \href{#method-remove_code}{\code{R6md$remove_code()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
view the private .char_code_blocks


Create a new R6md object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$new(rep_title, intro)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rep_title}}{Title of the report}

\item{\code{intro}}{Text to be included as introduction}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the private .char_code_blocks
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print2"></a>}}
\subsection{Method \code{print2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$print2(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_n_exp"></a>}}
\subsection{Method \code{set_n_exp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$set_n_exp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deparse_code"></a>}}
\subsection{Method \code{deparse_code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$deparse_code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_code"></a>}}
\subsection{Method \code{add_code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$add_code(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_code_block"></a>}}
\subsection{Method \code{add_code_block()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$add_code_block(expr, prefix = NULL, block_settings = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_code"></a>}}
\subsection{Method \code{run_code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$run_code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_code"></a>}}
\subsection{Method \code{remove_code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6md$remove_code()}\if{html}{\out{</div>}}
}

}
}
